+++ packages/spacecat-shared-data-access/src/models/preflight/index.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2026 Adobe. All rights reserved.
+ * Copyright 2025 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -10,21 +10,10 @@
  * governing permissions and limitations under the License.
  */
 
-const brandSemrushProjects = [
-  {
-    brandSemrushProjectId: '5b1e2d6a-7c5f-4a91-9b32-8e34ad11d201',
-    brandId: 'c3e1a4b6-2a8e-4d61-8b03-7d0a1d6b3201',
-    semrushProjectId: 'proj-fixture-us-en',
-    geoTargetId: 2840,
-    languageCode: 'en',
-  },
-  {
-    brandSemrushProjectId: '6c2f3e7b-8d6a-4b02-aa43-9f45be22e312',
-    brandId: 'c3e1a4b6-2a8e-4d61-8b03-7d0a1d6b3201',
-    semrushProjectId: 'proj-fixture-de-de',
-    geoTargetId: 2276,
-    languageCode: 'de',
-  },
-];
+import Preflight from './preflight.model.js';
+import PreflightCollection from './preflight.collection.js';
 
-export default brandSemrushProjects;
+export {
+  Preflight,
+  PreflightCollection,
+};